bigdecimal

+++ xsl_task_service/src/main/java/service/TaskAccount.java
@@ -1,7 +1,9 @@
 package service;
 
+import java.math.BigDecimal;
+
 public interface TaskAccount {
     int totalTask();
-    int totalMoney();
+    BigDecimal totalMoney();
     int totalDoneTask();
 }
